@sounisi5011/asciidoctor-extension-omit-line-breaksのコードをparse5を使用したものに統一
