+++ plugins/asciidoctor/extensions/omit-line-breaks/package.json
@@ -2,7 +2,6 @@
   "name": "@sounisi5011/asciidoctor-extension-omit-line-breaks",
   "version": "0.0.0",
   "private": true,
-  "main": "node/index.js",
   "dependencies": {
     "parse5": "^5.1.1"
   }
